+++ MoltenVK/MoltenVK/GPUObjects/MVKShaderModule.h
@@ -254,7 +254,8 @@ public:
 	 * If the Metal library compiler does not return within MVKConfiguration::metalCompileTimeout
 	 * nanoseconds, an error will be generated and logged, and nil will be returned.
 	 */
-	id<MTLLibrary> newMTLLibrary(NSString* mslSourceCode);
+	id<MTLLibrary> newMTLLibrary(NSString* mslSourceCode,
+								 const SPIRVToMSLConversionResults& shaderConversionResults);
 
 
 #pragma mark Construction